code/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_pytorch_select=0.2=gpu_0
absl-py=0.7.1=pypi_0
addict=2.2.0=pypi_0
apex=0.1=pypi_0
astor=0.7.1=pypi_0
astroid=2.2.5=py37_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
backports.csv=1.0.7=py_0
blas=1.0=mkl
bzip2=1.0.6=h14c3975_5
ca-certificates=2020.6.20=hecda079_0
cairo=1.16.0=h18b612c_1001
certifi=2020.6.20=py37hc8dfbb8_0
cffi=1.12.2=py37h2e261b9_1
chardet=3.0.4=py37hc8dfbb8_1006
cheroot=8.3.0=py37hc8dfbb8_1
cherrypy=18.5.0=py37_0
cloudpickle=0.8.1=py_0
cmake=3.14.4=pypi_0
cryptography=2.8=py37hb09aad4_2
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cxxfilt=0.2.1=py37h3340039_1
cycler=0.10.0=py37_0
cython=0.29.7=pypi_0
cytoolz=0.9.0.1=py37h14c3975_1
dask-core=1.2.0=py_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py37_1
dlib=19.17.0=pypi_0
docopt=0.6.2=pypi_0
expat=2.2.6=he6710b0_0
feedparser=5.2.1=py37_1
ffmpeg=4.1.3=h167e202_0
fontconfig=2.13.1=he4413a7_1000
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
future=0.18.2=py37hc8dfbb8_1
gast=0.2.2=pypi_0
gettext=0.19.8.1=hc5be6a0_1002
giflib=5.1.9=h516909a_0
glib=2.58.3=py37h6f030ca_1002
gmp=6.2.0=he1b5a44_0
gnutls=3.6.5=hd3a4fd2_1002
graphite2=1.3.13=h23475e2_0
grpcio=1.20.0=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.9.0=pypi_0
harfbuzz=2.4.0=h37c48d4_1
hdf5=1.10.5=nompi_h3c11f04_1104
icu=58.2=h9c2bf20_1
idna=2.8=pypi_0
importlib-metadata=1.6.1=py37hc8dfbb8_0
importlib_metadata=1.6.1=0
intel-openmp=2019.3=199
ipython=7.5.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
isort=4.3.19=py37_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.13.3=py37_0
joblib=0.14.1=py_0
jpeg=9c=h14c3975_1001
kiwisolver=1.0.1=py37hf484d3e_0
lame=3.100=h14c3975_1001
lazy-object-proxy=1.4.1=py37h7b6447c_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libiconv=1.15=h516909a_1005
liblapack=3.8.0=14_mkl
liblapacke=3.8.0=14_mkl
libopenblas=0.3.7=h5ec1e0e_6
libopus=1.3=h7b6447c_0
libpng=1.6.36=hbc83047_0
libprotobuf=3.12.3=h8b12597_0
libstdcxx-ng=8.2.0=hdf63c60_1
libsvm=323=0
libtiff=4.0.10=h2733197_2
libuuid=2.32.1=h14c3975_1000
libvpx=1.7.0=h439df22_0
libwebp=1.0.2=h576950b_1
libxcb=1.13=h1bed415_1
libxml2=2.9.9=he19cac6_0
libxslt=1.1.32=hae48121_1003
lxml=4.4.1=py37h7ec2d77_0
markdown=3.1=pypi_0
matplotlib=3.1.1=py37_0
matplotlib-base=3.1.1=py37hfd891ef_0
mccabe=0.6.1=py37_1
mkl=2019.4=243
mkl-service=2.3.0=py37h516909a_0
mkl_fft=1.0.13=py37h516909a_1
mkl_random=1.0.4=py37hf2d7682_0
mock=2.0.0=pypi_0
more-itertools=8.2.0=py_0
mysql-connector-c=6.1.11=hc9bbbdd_1005
mysqlclient=1.4.6=py37h3340039_1
ncurses=6.1=he6710b0_1
nettle=3.4.1=h1bed415_1002
ninja=1.9.0=pypi_0
numpy=1.17.3=py37h95a1406_0
numpy-base=1.18.1=py37hde5b4d6_1
olefile=0.46=py37_0
opencv=4.1.0=py37h3aa1047_6
opencv-contrib-python=4.1.0.25=pypi_0
opencv-python=4.1.0.25=pypi_0
openh264=1.8.0=hdbcaa40_1000
openssl=1.1.1g=h516909a_0
packaging=20.4=pyh9f0ad1d_0
pandas=0.24.2=py37he6710b0_0
parso=0.4.0=py_0
pattern=3.6.0=py_0
pbr=5.1.3=pypi_0
pcre=8.43=he6710b0_0
pdfminer.six=20191020=py_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=6.1.0=pypi_0
pip=19.0.3=py37_0
pixman=0.38.0=h7b6447c_0
pluggy=0.13.1=py37hc8dfbb8_2
portend=2.5=pyh9f0ad1d_2
progressbar2=3.39.3=pypi_0
prompt_toolkit=2.0.9=py37_0
protobuf=3.7.1=pypi_0
ptyprocess=0.6.0=py37_0
py=1.8.2=pyh9f0ad1d_0
pybind11=2.4.3=pypi_0
pycocotools=2.0=pypi_0
pycparser=2.19=py37_0
pycryptodome=3.9.7=py37hb05bdb2_1
pygments=2.4.0=py_0
pylint=2.3.1=py37_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.0=py_0
pysocks=1.7.1=py37hc8dfbb8_1
pyspellchecker=0.4.0=pypi_0
pytesseract=0.2.6=pypi_0
pytest=5.4.3=py37hc8dfbb8_0
python=3.7.3=h0371630_0
python-dateutil=2.8.0=py37_0
python-docx=0.8.10=py_0
python-utils=2.3.0=pypi_0
python_abi=3.7=1_cp37m
pytorch=1.3.1=cuda100py37h53c1284_0
pytube=9.5.0=pypi_0
pytz=2019.1=py_0
pywavelets=1.0.3=py37hdd07704_1
pyyaml=5.1=pypi_0
readline=7.0=h7b6447c_5
repoze.lru=0.7=py_0
requests=2.21.0=pypi_0
rope=0.14.0=py_0
routes=2.4.1=py37_0
scikit-learn=0.22.1=py37hcdab131_1
scipy=1.2.1=pypi_0
selenium=3.141.0=pypi_0
setuptools=41.0.0=py37_0
simplejson=3.17.0=py37h516909a_0
sip=4.19.8=pypi_0
six=1.12.0=py37_0
sortedcontainers=2.1.0=py_0
soupsieve=1.9.1=pypi_0
sqlite=3.27.2=h7b6447c_0
tempora=3.0.0=py_0
termcolor=1.1.0=pypi_0
tk=8.6.9=hed695b0_1003
toolz=0.9.0=py37_0
torchtext=0.3.1=pypi_0
torchvision=0.4.2=cuda100py37hecfc37a_0
tqdm=4.31.1=pypi_0
traitlets=4.3.2=py37_0
urllib3=1.24.3=pypi_0
wcwidth=0.1.7=py37_0
webvtt-py=0.4.2=pypi_0
werkzeug=0.15.2=pypi_0
wheel=0.33.1=py37_0
wordsegment=1.3.1=pypi_0
wrapt=1.11.1=py37h7b6447c_0
x264=1!152.20180806=h14c3975_0
xlrd=1.2.0=pypi_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.9=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.4=h14c3975_4
yacs=0.1.6=pypi_0
yaml=0.2.5=h516909a_0
zc.lockfile=2.0=py_0
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
